pyproject.toml
[tool.poetry]
name = "meteo-data-api"
version = "0.1.0"
description = ""
authors = ["guseva-sofya <guseva.sofya.p@gmail.com>"]
readme = "README.md"
packages = [{include = "meteo_data_api"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
psycopg2-binary = "^2.9.8"
alembic = "^1.12.0"
mypy = "^1.6.1"
types-psycopg2 = "^2.9.21.15"
pytest = "^7.4.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
